requirements.txt
streamlit
opencv-python-headless
numpy
Pillow
kagglehub
scikit-learn
matplotlib
